requirements.txt
colorama==0.4.6
google-generativeai==0.7.2